executor: split big test file into smaller files.
One big test file makes it hard to locate a specific test case.
